+++ common/src/main/java/net/modgarden/barricade/platform/BarricadePlatform.java
@@ -1,0 +1,6 @@
+package net.modgarden.barricade.platform;
+
+public enum BarricadePlatform {
+    FABRIC,
+    NEOFORGE
+}