+++ fabric/src/main/java/net/modgarden/barricade/fabric/client/platform/BarricadeClientPlatformHelperFabric.java
@@ -49,4 +49,9 @@ public class BarricadeClientPlatformHelperFabric implements BarricadeClientPlatf
         return new CreativeOnlyBakedModel(model, requiredItem);
     }
 
+    @Override
+    public String getPackName(String name) {
+        return name;
+    }
+
 }