+++ neoforge/src/main/java/net/modgarden/barricade/neoforge/client/platform/BarricadeClientPlatformHelperNeoForge.java
@@ -53,4 +53,9 @@ public class BarricadeClientPlatformHelperNeoForge implements BarricadeClientPla
         return new CreativeOnlyBakedModel(model, requiredItem);
     }
 
+    @Override
+    public String getPackName(String name) {
+        return "mod/" + name;
+    }
+
 }